# backend/backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Movie Backend API",
    "description": "An API for movie management",
    "version": "v1"
  },
  "paths": {
    "/api/Movies/popular-movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Movies/top-rated": {
      "get": {
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Movies/upcoming-movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Movies/{id}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Movies/add-to-offline": {
      "post": {
        "tags": [
          "Movies"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddToOfflineListRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Movies/offline-list": {
      "get": {
        "tags": [
          "Movies"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Movies/remove-from-offline": {
      "delete": {
        "tags": [
          "Movies"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RemoveFromOfflineList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "AddToOfflineListRequest": {
      "type": "object",
      "properties": {
        "movieId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "RemoveFromOfflineList": {
      "type": "object",
      "properties": {
        "movieId": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Please insert JWT with Bearer into field"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}